// src/Engine/GUI/Checkbox.d.ts
import Vec2 from "../Maths/Vec2";
import GuiObject from "./GuiObject";
export default class Checkbox extends GuiObject {
    position: Vec2;
    textSize: number;
    private inputNode;
    private label;
    constructor();
    getElement(): HTMLDivElement;
    getInputElement(): HTMLInputElement;
    getChecked(): boolean;
    draw(): void;
}